+++ src/main/java/org/rivierarobotics/lib/MotorUtil.java
@@ -56,7 +56,7 @@ public class MotorUtil {
      *
      * @since 0.1.0
      */
-    public static void setupMotionMagic(FeedbackDevice sensor, PIDConfig pidConfig, int maxVel, BaseTalon... motors) {
+    public static void setupMotionMagic(FeedbackDevice sensor, PIDConfig pidConfig, double maxVel, BaseTalon... motors) {
         int periodMs = 10;
         int timeoutMs = 10;
         for (BaseTalon motor : motors) {
@@ -100,7 +100,7 @@ public class MotorUtil {
      * @param motors the motors to apply the soft limits onto.
      * @since 0.1.0
      */
-    public static void setSoftLimits(int forward, int reverse, BaseTalon... motors) {
+    public static void setSoftLimits(double forward, double reverse, BaseTalon... motors) {
         for (BaseTalon motor : motors) {
             motor.configForwardLimitSwitchSource(LimitSwitchSource.Deactivated, LimitSwitchNormal.NormallyOpen);
             motor.configReverseLimitSwitchSource(LimitSwitchSource.Deactivated, LimitSwitchNormal.NormallyOpen);
